Extract parameter resolution in own class
-All parameter handling logic extracted from Main class into a dedicated ParameterResolver
-Improved error handling - dedicated parameter error messages
-Refactor - addressed code duplication
