+++ lang/frontend/cli/src/main/java/tools/vitruv/neojoin/cli/ParameterResolutionException.java
@@ -1,0 +1,12 @@
+package tools.vitruv.neojoin.cli;
+
+class ParameterResolutionException extends RuntimeException {
+
+    ParameterResolutionException(String message) {
+        super(message);
+    }
+
+    ParameterResolutionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
